Separated state and ui
