--- src/ui/mod.rs
@@ -1,3 +1,0 @@
-mod app;
-
-pub use app::App;
